refactor: Seperate Test code
